[TouchRunner] Restore the original code path when not wrapping NUnit xml output.
Asking NUnit to write xml into a StringWriter and then forwarding that to a
TextWriter is troublesome, because StringWriter identifies as UTF16 (and it
can't be changed without subclassing), while the TextWriter can use any other
encoding. The problem arises because NUnit will write an xml declaration, and
when writing to a StringWriter the xml declaration will say its encoding is
UTF16, thus causing trouble later on when we write that output into a UTF8
text writer.

Solution: restore the original code path when not wrapping NUnit xml output,
so that we ask NUnit to write directly to the UTF8 text writer.

When wrapping the NUnit xml output, we already strip off the xml declaration,
so that's not problem in the first place.