[TCP] Add support to use a tcp tunnel.
Using a tcp tunnle inverts the way we communicate with the device. We
move from a TcpClient to a TcpListener. In this case we need to wait
until we have a client connected that sends us a ping to start running
the tests.

This commit does not mean that we won't be able to use a normal tcp
conection, just means that if required, we can use a tunnel