[TouchRunner] Only update the UI for a single test after the test has finished running.
Only update the UI for a single test after we're notified that the test has
finished running. This is a step towards making async tests work properly,
where we can't assume that the test has finished after calling Run.

Also we only want to show more details for a test if we ran that test as a
result of tapping on it.