[TouchRunner] Try to make MacRunner exit nicely. (#100)
Try to exit the NSApplication run loop nicely, by calling NSApplication.Stop.
The problem is that there seems to be certain requirements for stopping the
NSApplication: one being that apparently there must be a window around. So
create a window (that shows up in the background). Then at the end we have to
post something to the event loop after calling Stop (because calling Stop
apparently only sets a flag that's checked after processing an event). And
just in case seomthing goes wrong, try two different ways of posting events to
the event loop.